move OrientedSource to source.image.py
