+++ src/aspire/abinitio/__init__.py
@@ -4,6 +4,3 @@ from .sync_voting import SyncVotingMixin
 # isort: off
 from .commonline_sync import CLSyncVoting
 from .commonline_c3_c4 import CLSymmetryC3C4
-
-# isort: on
-from .orientation_src import OrientEstSource
