+++ src/aspire/source/__init__.py
@@ -1,7 +1,12 @@
 import logging
 
 from aspire.source.coordinates import BoxesCoordinateSource, CentersCoordinateSource
-from aspire.source.image import ArrayImageSource, ImageSource, IndexedSource
+from aspire.source.image import (
+    ArrayImageSource,
+    ImageSource,
+    IndexedSource,
+    OrientedSource,
+)
 from aspire.source.relion import RelionSource
 from aspire.source.simulation import Simulation
